# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Gateway Registry",
    "description": "A registry and management system for Model Context Protocol (MCP) servers",
    "version": "1.0.11"
  },
  "paths": {
...
    "/api/management/iam/users": {
      "get": {
        "tags": [
          "Management API"
        ],
        "summary": "Management List Users",
        "description": "List users from the configured identity provider (admin only).",
        "operationId": "management_list_users_api_management_iam_users_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/management/iam/users/m2m": {
      "post": {
        "tags": [
          "Management API"
        ],
        "summary": "Management Create M2M User",
        "description": "Create a service account client and return its credentials (admin only).",
        "operationId": "management_create_m2m_user_api_management_iam_users_m2m_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/M2MAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/management/iam/users/human": {
      "post": {
        "tags": [
          "Management API"
        ],
        "summary": "Management Create Human User",
        "description": "Create a human user and assign groups (admin only).",
        "operationId": "management_create_human_user_api_management_iam_users_human_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/management/iam/users/{username}": {
      "delete": {
        "tags": [
          "Management API"
        ],
        "summary": "Management Delete User",
        "description": "Delete a user by username (admin only).",
        "operationId": "management_delete_user_api_management_iam_users__username__delete",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/management/iam/groups": {
      "get": {
        "tags": [
          "Management API"
        ],
        "summary": "Management List Groups",
        "description": "List IAM groups from the configured identity provider (admin only).",
        "operationId": "management_list_groups_api_management_iam_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Management API"
        ],
        "summary": "Management Create Group",
        "description": "Create a new group in the identity provider and MongoDB (admin only).\n\nThis creates the group in both:\n1. The configured identity provider (Keycloak or Entra ID)\n2. MongoDB scopes collection for authorization",
        "operationId": "management_create_group_api_management_iam_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/management/iam/groups/{group_name}": {
      "delete": {
        "tags": [
          "Management API"
        ],
        "summary": "Management Delete Group",
        "description": "Delete a group from the identity provider and MongoDB (admin only).\n\nThis deletes the group from both:\n1. The configured identity provider (Keycloak or Entra ID)\n2. MongoDB scopes collection",
        "operationId": "management_delete_group_api_management_iam_groups__group_name__delete",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GroupCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupCreateRequest",
        "description": "Payload for creating a Keycloak group."
      },
      "GroupDeleteResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupDeleteResponse",
        "description": "Response when a Keycloak group is deleted."
      },
      "GroupListResponse": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupSummary"
            },
            "type": "array",
            "title": "Groups"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "GroupListResponse",
        "description": "Response for listing groups."
      },
      "GroupSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path"
        ],
        "title": "GroupSummary",
        "description": "Group information."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HumanUserRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "firstname": {
            "type": "string",
            "minLength": 1,
            "title": "Firstname"
          },
          "lastname": {
            "type": "string",
            "minLength": 1,
            "title": "Lastname"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Groups"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Initial password (optional, generated elsewhere)"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "firstname",
          "lastname",
          "groups"
        ],
        "title": "HumanUserRequest",
        "description": "Payload for creating a Keycloak human user."
      },
      "M2MAccountRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Groups"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "groups"
        ],
        "title": "M2MAccountRequest",
        "description": "Payload for creating a Keycloak service account client."
      },
...
      "UserDeleteResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "UserDeleteResponse",
        "description": "Standard response returned when a Keycloak user is deleted."
      },
      "UserListResponse": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserSummary"
            },
            "type": "array",
            "title": "Users"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "UserListResponse",
        "description": "Wrapper for list users endpoint."
      },
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "firstName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "lastName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastname"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "UserSummary",
        "description": "Subset of user information exposed through the API."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token obtained from Keycloak OAuth2 authentication. Include in Authorization header as: `Authorization: Bearer <token>`"
      }
    }
  }
}